+++ airportsCDT.h
@@ -15,7 +15,7 @@ typedef struct destination * tDestination;
 enum days {SUN=0, MON, TUE, WED, THU, FRI, SAT};
 
 
-struct flight 
+struct flight
 {
     char date [ 11 ];		/* Formato DD/MM/YYYY */
     char time [ 6 ];		/* Formato HH:MM */
@@ -33,8 +33,8 @@ struct airport
 	char * name;
 	char iata[4];
 	int isInternational;
-	size_t takeOffs;
-	size_t landings;
+	unsigned int takeOffs;
+	unsigned int landings;
 	enum days day;
 	destinationADT destinations;	/* Lista de struct destination */
 };
@@ -42,14 +42,14 @@ struct airport
 struct destination
 {
 	char * oaci; /* debo poner char * por los codigos del tipo AR-XXXX sino seria char dest[5]*/
-	size_t landings;
-	size_t takeOffs;
+	unsigned int landings;
+	unsigned int takeOffs;
 };
 
 struct airline
 {
 	char * name;
-	size_t movs;
+	unsigned int movs;
 };
 
 #endif
